requirements.txt
asgiref==3.2.10
Django==3.1
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-richtextfield==1.6.1
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
